% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPrefix-methods.R
\name{addPrefix}
\alias{addPrefix}
\alias{addPrefix,GRanges-method}
\title{Add prefixes to sequence names of a \code{GRanges} object}
\usage{
addPrefix(x, prefix)

\S4method{addPrefix}{GRanges}(x, prefix)
}
\arguments{
\item{x}{\code{GRanges} object}

\item{prefix}{\code{character()} The prefixes to prepend. Must be either a
single string or a character vector with length equal to the length of
\code{x}.}
}
\value{
\code{GRanges} object with the modified sequence names. Only the
  ranges, strands, and metadata columns will be retained from \code{x}.
}
\description{
Add prefixes to sequence names of a \code{GRanges} object
}
\seealso{
\code{\link{rmPrefix}}
}
