% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCoverage-methods.R
\name{cnvCoverage}
\alias{cnvCoverage}
\alias{cnvCoverage,GRanges,GRanges-method}
\title{Compute coverage between two \code{GRanges} objects}
\usage{
cnvCoverage(x, y)

\S4method{cnvCoverage}{GRanges,GRanges}(x, y)
}
\arguments{
\item{x, y}{\code{GRanges} objects. Because the intent is to compute
coverage of CNVs, both objects should be stranded according to SV type.
"DUP"s should be "+" and "DEL"s should be "-".}
}
\value{
A \code{double} vector equal to the length of \code{x} with the
  coverage of each range in \code{x}.
}
\description{
For every range in \code{x}, compute the proportion of the range that is
overlapped by ranges in \code{y}. This function is intended to be used to
get the coverage between sets of CNVs.
}
