% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_gr.R
\name{df_to_gr}
\alias{df_to_gr}
\alias{df_to_gr.data.frame}
\alias{df_to_gr.data.table}
\title{Create a \code{GRanges} object from a \code{data.frame}}
\usage{
df_to_gr(x, ...)

\method{df_to_gr}{data.frame}(x, cnv = FALSE, ...)

\method{df_to_gr}{data.table}(x, cnv = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.frame} with at least the columns \code{chr},
\code{start}, and \code{end}. A \code{strand} column will be used for
strand if available and \code{cnv} is \code{FALSE}. All columns that are
not \code{chr}, \code{start}, \code{end}, or otherwise prohibited will
become metadata columns.}

\item{...}{Other arguments passed to methods.}

\item{cnv}{If \code{TRUE}, the genomic ranges will be treated as CNVs. In
this case, there should be an \code{svtype} column in \code{x} indicating
the CNV type (either 'DUP' or 'DEL'). DUPs will be given a strand of '+'
and DELs will given a strand of '-'.}
}
\value{
A \code{GRanges} object.
}
\description{
Create a \code{GRanges} object from a \code{data.frame} containing genomic
ranges.
}
\details{
The columns 'seqnames', 'ranges', 'strand', 'seqlevels', 'seqlengths',
'isCircular', 'start', 'end', 'width', and 'element' are not allowed names
for a \code{GRanges} metadata column so columns with these names will be
removed from \code{x} with a warning before assigning them to metadata
columns.
}
