% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_region_by_bins.R
\name{expand_region_by_bins}
\alias{expand_region_by_bins}
\title{Expand a genomic region based on overlap of genomic bins}
\usage{
expand_region_by_bins(region, bins, pad = 0.2)
}
\arguments{
\item{region}{A \code{gregion} object.}

\item{bins}{A \code{GBins} object.}

\item{pad}{The fraction of overlapped bins to use to expand \code{region}.}
}
\value{
A \code{gregion} object representing the expanded region. If
  \code{region} does not overlap any bins, it will be returned unmodified.
}
\description{
Find the genomic bins that a genomic region overlaps and expand the region
by a proportion of the overlapped bins.
}
\details{
If \code{gregion} overlaps \var{N} bins in \code{bins},
then \eqn{\lceil N \times pad \rceil}{ceil(N * pad)} bins will be added
upstream and downstream of the overlapping set and the total region covered
by the bins will become the new expanded region.
}
\seealso{
expandRangesByBins
}
