% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_dcr.R
\name{get_batch_dcr}
\alias{get_batch_dcr}
\title{Retrieve a dCR region for a gCNV batch}
\usage{
get_batch_dcr(x, paths, bound_dcr = TRUE)
}
\arguments{
\item{x}{A \code{gregion} object for the target region.}

\item{paths}{A character vector of paths. The length must be at least 1.}

\item{bound_dcr}{If \code{TRUE}, all dCR values less than 0 will be set to
0 and all dCR values greater than 5 will be set to 5. Otherwise the dCR
matrix will be left as is.}
}
\value{
A \code{data.table} with the portion of the dCR matrix overlapping
  the query region. The first three columns will be \code{'chr'},
  \code{'start'}, and \code{'end'} corresponding to the dCR regions. The
  remaining columns will be the samples and their dCR values.
}
\description{
Retrieve all the dCR matrices for a single gCNV batch, subset to the ranges
overlapping a given genomic region.
}
\details{
It is critical that all of the dCR files passed to this function are actually
for the same batch because the ranges extracted from each matrix are
assumed to be exactly the same and in the same order (which is the case for
matrices from the same batch). This function will only check that the
number, not the order or coordinates, of the extracted regions from each
sub-matrix is the same.
}
