% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_dcr.R
\name{get_samples_dcr}
\alias{get_samples_dcr}
\title{Retrieve a dCR region for samples in a gCNV batch}
\usage{
get_samples_dcr(x, paths, samples, include_bg = FALSE, ...)
}
\arguments{
\item{x}{A \code{gregion} object for the target region.}

\item{paths}{A character vector of paths. The length must be at least 1.}

\item{samples}{A character vector of sample IDs. Must contain at least one
ID and all samples must be from the same batch.}

\item{include_bg}{If \code{TRUE}, all of the other samples not among
\code{samples} will be included in the returned \code{data.table}.
Otherwise, only the requested samples are returned.}

\item{...}{Any additional arguments to pass to \code{\link{get_batch_dcr}}.}
}
\value{
A \code{data.frame} with the portion of the dCR matrix overlapping
  the query region. The first three columns will be \code{'chr'},
  \code{'start'}, and \code{'end'}, corresponding to the dCR ranges. The
  next \var{N} columns (where \var{N} is the number samples requested) will
  be the dCR columns for the requested samples. The order of these columns is
  not guaranteed. The remaining columns will be the remaining samples in the
  batch, if requested. If any sample ID in \code{samples} is not found in
  the dCR matrix, this function will signal an error.
}
\description{
Retrieve the dCR matrices for samples in a single gCNV batch, subset to the
ranges overlapping a given genomic region.
}
\seealso{
get_batch_dcr
}
