% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregion.R
\name{gregion}
\alias{gregion}
\title{Create a \code{gregion} object}
\usage{
gregion(chr, start, end)
}
\arguments{
\item{chr}{Name of the sequence.}

\item{start}{Start of the region. Should be 1-based.}

\item{end}{End of the region. Should be closed.}
}
\value{
A new \code{gregion} object.
}
\description{
Create a \code{gregion} object.
}
\details{
A \code{gregion} object represents a single genomic region. It is useful for
working with genomic coordinates as a single object because it is named list
with \code{chr}, \code{start}, and \code{end} components.
}
