% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logging}
\alias{logging}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\title{Log to stderr}
\usage{
log_info(msg)

log_warn(msg)

log_error(msg)
}
\arguments{
\item{msg}{A character vector of the message to log. \code{msg} can also be
an R object for which \code{as.character()} is defined and works, in which
case the result of \code{as.character(msg)} will be used as the message.
If \code{msg} is a string without any newlines, the message will be
written as a single line. If \code{msg} contains newlines or has a length
greater than one, then the message will be written as a block. See
details.}
}
\description{
Write logging messages of different levels to stderr.
}
\details{
These functions enable logging messages of different levels (i.e. INFO,
WARN, and ERROR) to stderr without using signals. The different levels
refers to what tags the messages will have and not in the sense of
application verbosity or directing messages of different levels to different
places. Every message will be written to stderr with a string prefix
consisting of the current date, time and log level tag. In addition to the
different log levels, there are two different forms of log messages: single
line and block. A single line message will be written on a single line. A
block comprises a opening tag, the message to log, and a closing tag. The
opening tag is a single line consisting of the usual prefix followed by the
string "LOGBLOCK". The closing tag is a single line consisting of the string
"LOGBLOCK".
}
