% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcnv_bins.R
\name{read_gcnv_bins}
\alias{read_gcnv_bins}
\title{Read gCNV bins file}
\usage{
read_gcnv_bins(path, reduce = FALSE)
}
\arguments{
\item{path}{Path to the file.}

\item{reduce}{Should overlapping ranges be merged?}
}
\value{
A \code{GBins} object.
}
\description{
Read the file of genomic ranges used by gCNV to bin the genome.
}
\details{
The file should be a three-column tab-delimited file containing the
chromosome, start, and end of each interval, in that order. Coordinates are
assumed to be 1-start, closed. Any columns beyond the first three will be
ignored. The intervals must be disjoint.
}
