% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmPrefix-methods.R
\name{rmPrefix}
\alias{rmPrefix}
\alias{rmPrefix,GRanges-method}
\title{Remove prefixes from sequence names of a \code{GRanges} object}
\usage{
rmPrefix(x, prefix)

\S4method{rmPrefix}{GRanges}(x, prefix)
}
\arguments{
\item{x}{\code{GRanges} object}

\item{prefix}{\code{character()} The prefixes to prepend. Must be either a
single string or a character vector with length equal to the length of
\code{x}. \code{NA} prefixes are not allowed.}
}
\value{
\code{GRanges} object with the modified sequence names. Only the
  ranges, strands, and metadata columns will be retained from \code{x}. If
  removing a prefix would result in an empty string, an error will be
  signalled.
}
\description{
Remove prefixes from sequence names of a \code{GRanges} object
}
\seealso{
\code{\link{addPrefix}}
}
