% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregion.R
\name{to_string}
\alias{to_string}
\title{Create a string representation of an object.}
\usage{
to_string(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A string representing the object.
}
\description{
S3 generic to create strings from objects.
}
\details{
Methods should accept a single object and return a single string (character
vector of length 1).
}
