% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBins-class.R
\name{toBinSpace}
\alias{toBinSpace}
\alias{toBinSpace,GRanges,GBins-method}
\title{Map genomic ranges to genomic bins}
\usage{
toBinSpace(x, y)

\S4method{toBinSpace}{GRanges,GBins}(x, y)
}
\arguments{
\item{x}{A \code{GRanges} object.}

\item{y}{A \code{GBins} object.}
}
\value{
A \code{GRanges} object with mapped coordinates.
}
\description{
Convert a set of genomic ranges to coordinates defined by spans of genomic
bins.
}
\details{
`toBinSpace` converts each genomic bin to its relative position (first bin
is 1, second bin is 2, and so on) and then maps each genomic range to its
span of bins. Thus, if a genomic range overlapped bins 7 through 13, its
coordinates would become 7 to 13. Strand is ignored when determining
overlap, but sequence name is respected.
}
