% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_gr.R
\name{df_to_gr}
\alias{df_to_gr}
\title{Make a \code{GRanges} object from a \code{data.frame}}
\usage{
df_to_gr(x)
}
\arguments{
\item{x}{A \code{data.frame} with at least the columns "chr", "start", and
"end". A "strand" column will be used for strand if available. All other
columns will become the metadata columns.}
}
\value{
A \code{GRanges} object.
}
\description{
Make a \code{GRanges} object from a \code{data.frame}
}
