% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dcr_list.R
\name{read_dcr_list}
\alias{read_dcr_list}
\title{Read list of dCR matrix paths}
\usage{
read_dcr_list(path)
}
\arguments{
\item{path}{Path to the file of dCR matrix paths. There should be one dCR
matrix path per line.}
}
\value{
\code{hashtab} in which the keys are batch IDs and the values are
  paths to the dCR matrices for the batch.
}
\description{
Read a list of dCR matrix paths and create a hash table from gCNV batch to
dCR matrix path.
}
\details{
The paths must be of the form
"path/to/\{BATCH\}_((CASE)|(COHORT)).dcr.bed.gz" which is the default for
gCNV. The CASE and COHORT files will be combined for a batch such that
querying the hash table for a batch will return a character vector with the
paths to both the CASE and COHORT matrices, if they are available.
}
