% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pedigree.R
\name{read_pedigree}
\alias{read_pedigree}
\title{Read a pedigree file}
\usage{
read_pedigree(path)
}
\arguments{
\item{path}{Path to the pedigree file.}
}
\value{
\code{data.table}.
}
\description{
A read a pedigree file and peform minor cleaning.
}
\details{
The file must be tab-delimited with six columns in the following order:
\enumerate{
  \item family_id
  \item sample_id
  \item paternal_id
  \item maternal_id
  \item sex
  \item phenotype
}
Only the phenotypes -9, 0, 1, and 2 are accepted with -9 and 0 meaning
unknown, 1 meaning unaffected and 2 meaning affected. All other phenotypes
will be silently converted to 0 and then -9 and 0 will be converted to `NA`.
Sex can be 1 for male and 2 for female. All other sex codes will be
converted to `NA`.  Any empty or duplicate sample IDs will be silently
removed. It is better for the user to properly sanitize the file than to
rely on this function to perform the cleaning.
}
